+++ bin/check-marathon-task.rb
@@ -47,10 +47,32 @@ require 'json'
 class MarathonTaskCheck < Sensu::Plugin::Check::CLI
   check_name 'CheckMarathonTask'
 
-  option :server, short: '-s SERVER', long: '--server SERVER', required: true
-  option :port, short: '-p PORT', long: '--port PORT', default: 8080
-  option :task, short: '-t TASK', long: '--task TASK', required: true
-  option :instances, short: '-i INSTANCES', long: '--instances INSTANCES', required: true, proc: proc(&:to_i)
+  option :server,
+         short: '-s SERVER',
+         long: '--server SERVER',
+         required: true
+
+  option :port,
+         short: '-p PORT',
+         long: '--port PORT',
+         default: 8080
+
+  option :uri,
+         description: 'Endpoint URI',
+         short: '-u URI',
+         long: '--uri URI',
+         default: '/v2/tasks?status=running'
+
+  option :task,
+         short: '-t TASK',
+         long: '--task TASK',
+         required: true
+
+  option :instances,
+         short: '-i INSTANCES',
+         long: '--instances INSTANCES',
+         required: true,
+         proc: proc(&:to_i)
 
   def run
     if config[:instances].zero?
@@ -58,9 +80,10 @@ class MarathonTaskCheck < Sensu::Plugin::Check::CLI
     end
 
     failures = []
+    uri = config[:uri]
     config[:server].split(',').each do |s|
       begin
-        url = URI.parse("http://#{s}:#{config[:port]}/v2/tasks?status=running")
+        url = URI.parse("http://#{s}:#{config[:port]}#{uri}")
         req = Net::HTTP::Get.new(url)
         req.add_field('Accept', 'application/json')
         r = Net::HTTP.new(url.host, url.port).start do |h|