+++ bin/check-mesos.rb
@@ -30,8 +30,8 @@ require 'sensu-plugin/check/cli'
 require 'rest-client'
 
 # Mesos default ports are defined here: http://mesos.apache.org/documentation/latest/configuration
-MASTER_DEFAULT_PORT = '5050'.freeze
-SLAVE_DEFAULT_PORT = '5051'.freeze
+MASTER_DEFAULT_PORT ||= '5050'.freeze
+SLAVE_DEFAULT_PORT ||= '5051'.freeze
 
 class MesosNodeStatus < Sensu::Plugin::Check::CLI
   option :server,
@@ -52,6 +52,12 @@ class MesosNodeStatus < Sensu::Plugin::Check::CLI
          long: '--port PORT',
          required: false
 
+  option :uri,
+         description: 'Endpoint URI',
+         short: '-u URI',
+         long: '--uri URI',
+         default: '/health'
+
   option :timeout,
          description: 'timeout in seconds',
          short: '-t TIMEOUT',
@@ -62,20 +68,19 @@ class MesosNodeStatus < Sensu::Plugin::Check::CLI
   def run
     mode = config[:mode]
     servers = config[:server]
+    uri = config[:uri]
     case mode
     when 'master'
       port = config[:port] || MASTER_DEFAULT_PORT
-      uri = '/master/health'
     when 'slave'
       port = config[:port] || SLAVE_DEFAULT_PORT
-      uri = '/slave(1)/health'
     end
     failures = []
     servers.split(',').each do |server|
       begin
         r = RestClient::Resource.new("http://#{server}:#{port}#{uri}", timeout: config[:timeout]).get
         if r.code != 200
-          failures << "#{config[:mode]} on #{server} is not responding"
+          failures << "#{mode} on #{server} is not responding"
         end
       rescue Errno::ECONNREFUSED, RestClient::ResourceNotFound, SocketError
         failures << "Mesos #{mode} on #{server} is not responding"