+++ CHANGELOG.md
@@ -4,8 +4,16 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This CHANGELOG follows the format listed at [Keep A Changelog](http://keepachangelog.com/)
 
 ## [Unreleased]
+
+### Added
+- check-metronome.rb: Check if Metronome is running
+
+### Changed
 - check-marathon-task.rb: Use the health check results to verify that a task is running.
 - check-marathon-task.rb: Rename incorrect "state" parameter to "status".
+- All checks now have a configurable API endpoint using --uri or -u
+- Support the latest Mesos API
+- Dropped support for Ruby 1.9.3
 
 ## [0.1.1] - 2016-03-04
 ### Added